Just move OriginalRoad to its own module
